tests/wpt/tests/css/css-scroll-snap-2/scroll-start-target/scroll-start-target-span.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title> CSS Scroll Snap 2 Test: scroll-start-target with a &lt;span> element</title>
    <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#scroll-start-target">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/css-animations/support/testcommon.js"></script>
  </head>
  <body>
    <style>
      * {
        margin: 0px;
      }
      .space {
        width: 150%;
        height: 150%;
      }
      .box {
        height: 50px;
        width: 50px;
        border: solid 1px black;
        background-color: turquoise;
      }
      .target {
        scroll-start-target: auto;
      }
      .scroller {
        overflow: scroll;
        height: 100px;
        width: 100px;
      }
    </style>
    <div id="scroller" class="scroller">
      <div class="space"></div>
      <span id="target" class="target">
        <div class="box"></div>
      </span>
      <div class="space"></div>
    </div>
    <script>
      promise_test(async (t) => {
        await waitForAnimationFrames(2);
        const scroller = document.getElementById("scroller");
        const target = document.getElementById("target");

        assert_equals(scroller.scrollTop, target.offsetTop,
          "scroller is initially scrolled to it <span> scroll-start-target");
      }, "<span> scroll-start-target is honored");
    </script>
  </body>
</html>
